feat: extract ic update user
